requirements.txt
django-autoslug
